build: make rewriting optional, but default, with `--no-rewrite` flag

+++ src/build/mod.rs
@@ -1158,6 +1158,7 @@ async fn fetch_dependencies(
     default_world: Option<&str>,
     include: &HashSet<PathBuf>,
     exclude: &HashSet<PathBuf>,
+    no_rewrite: bool,
     force: bool,
     verbose: bool,
 ) -> Result<()> {
@@ -1174,6 +1175,7 @@ async fn fetch_dependencies(
         default_world,
         vec![], // TODO: what about deps-of-deps?
         vec![],
+        no_rewrite,
         false,
         force,
         verbose,
@@ -1210,6 +1212,7 @@ async fn fetch_dependencies(
             default_world,
             local_dep_deps,
             vec![],
+            no_rewrite,
             false,
             force,
             verbose,
@@ -1525,6 +1528,7 @@ async fn compile_package(
     add_paths_to_api: &Vec<PathBuf>,
     include: &HashSet<PathBuf>,
     exclude: &HashSet<PathBuf>,
+    no_rewrite: bool,
     force: bool,
     verbose: bool,
     ignore_deps: bool, // for internal use; may cause problems when adding recursive deps
@@ -1547,6 +1551,7 @@ async fn compile_package(
             default_world,
             include,
             exclude,
+            no_rewrite,
             force,
             verbose,
         )
@@ -1654,6 +1659,7 @@ pub async fn execute(
     default_world: Option<&str>,
     local_dependencies: Vec<PathBuf>,
     add_paths_to_api: Vec<PathBuf>,
+    no_rewrite: bool,
     reproducible: bool,
     force: bool,
     verbose: bool,
@@ -1737,9 +1743,16 @@ pub async fn execute(
 
     check_process_lib_version(&package_dir.join("Cargo.toml"))?;
 
-    let rewritten_dir = copy_and_rewrite_package(package_dir)?;
+    // live_dir is the "dir that is being built" or is "live";
+    //  if `no_rewrite`, that is just `package_dir`;
+    //  else, it is the modified copy that is in `target/rewrite/`
+    let live_dir = if no_rewrite {
+        PathBuf::from(package_dir)
+    } else {
+        copy_and_rewrite_package(package_dir)?
+    };
 
-    let ui_dirs = get_ui_dirs(&rewritten_dir, &include, &exclude)?;
+    let ui_dirs = get_ui_dirs(&live_dir, &include, &exclude)?;
     if !no_ui && !ui_dirs.is_empty() {
         if !skip_deps_check {
             let mut recv_kill = make_fake_kill_chan();
@@ -1754,7 +1767,7 @@ pub async fn execute(
 
     if !ui_only {
         compile_package(
-            &rewritten_dir,
+            &live_dir,
             skip_deps_check,
             features,
             url,
@@ -1764,6 +1777,7 @@ pub async fn execute(
             &add_paths_to_api,
             &include,
             &exclude,
+            no_rewrite,
             force,
             verbose,
             ignore_deps,
@@ -1771,10 +1785,12 @@ pub async fn execute(
         .await?;
     }
 
-    if package_dir.join("pkg").exists() {
-        fs::remove_dir_all(package_dir.join("pkg"))?;
+    if !no_rewrite {
+        if package_dir.join("pkg").exists() {
+            fs::remove_dir_all(package_dir.join("pkg"))?;
+        }
+        copy_dir(live_dir.join("pkg"), package_dir.join("pkg"))?;
     }
-    copy_dir(rewritten_dir.join("pkg"), package_dir.join("pkg"))?;
 
     let metadata = read_metadata(package_dir)?;
     let pkg_publisher = make_pkg_publisher(&metadata);
